[PE] Fix layer under-invalidation when visiblity status changes
https://chromium-review.googlesource.com/c/chromium/src/+/1017917
added code to invalidate the paint layer when its
PaintedOutputInvisible() status changes. The problem was that the
function doesn't merely depend on the old style or the new style,
but PaintsWithTransparency() based on the new style and the current
compositing status, so PaintedOutputInvisible(old_style) was not
the actual old invisible status. This caused we missed repaint of
the layer in the following case:

frame 1:
  opacity 0 (we don't paint the layer)
frame 2:
  opacity 0 -> 1 (PaintedOutputInvisible(old_style) was wrong, so
                  we failed to detect the change of it)
  layout and compositing update [1]
  opacity 1 -> 0.9

[1] If we also do prepaint here, change of existence of the effect
    node will trigger repaint of the layer, so the issue reproduces
    only in the above sequence (tested in the new wpt test).

Now let PaintedOutputInvisible() only access the input style.

Bug: 922084
Change-Id: Iec98b0304275cff968ed8053776f7c9a77a9625c