Adds pre shared key support to RTCQuicTransport.
This change adds the PSK APIs to the RTCQuicTransport object. This is
done with a readonly key, connect() and listen() methods. This change
keeps the C++ objects certificate support with the current APIs, but
removes these from being exposed in the web idl file. Tests are updated
by moving web platform tests that involve certificates into C++
tests.

Bug:874296
Change-Id: I58500dca36bfdb6fdb605d540ccbfc49ed6ec878